Move new group CLI into a separate file
